move gas tests where they belong
